+++ src/main/java/model/heuristic/Heuristic.java
@@ -3,5 +3,5 @@ package model.heuristic;
 import model.State;
 
 public interface Heuristic {
-    int getValue(State state);
+    double getValue(State state);
 }
