+++ src/main/java/model/heuristic/nn/LocationOfHighValueBlocks.java
@@ -5,7 +5,7 @@ import model.heuristic.Heuristic;
 
 public class LocationOfHighValueBlocks implements Heuristic {
     @Override
-    public int getValue(State state) {
+    public double getValue(State state) {
         int[][] board = state.getBoard();
         int sum = 0;
